Made database functions into submodule
