+++ gens/db/__init__.py
@@ -1,0 +1,2 @@
+from .db import init_database_connection as init_database
+from .db import query_variants, query_records_in_region
